Fixed throttle Effect operator implementation
