+++ Package.swift
@@ -5,7 +5,7 @@ import PackageDescription
 let package = Package(
   name: "rxswift-composable-architecture",
   platforms: [
-    .iOS(.v12),
+    .iOS(.v13),
     .macOS(.v10_15),
     .tvOS(.v13),
     .watchOS(.v6),
